+++ src/main/java/com/by/model/GiftCoupon.java
@@ -19,7 +19,7 @@ public class GiftCoupon extends Coupon {
         super();
     }
 
-    public GiftCoupon(Long id) {
+    public GiftCoupon(int id) {
         setId(id);
     }
